requirements.txt
numpy==1.21.5
opencv-python==4.5.5.62
pandas==1.3.5
Pillow==9.3.0
pyradiomics==3.0.1
PyWavelets==1.3.0
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
SimpleITK==2.1.1
